+++ src/main/java/com/nylas/services/Tunnel.java
@@ -58,13 +58,12 @@ public class Tunnel extends WebSocketClient {
 	 */
 	@Override
 	public void onOpen(ServerHandshake handshakedata) {
-		log.trace("Opening websocket connection");
 		webhookHandler.onOpen(handshakedata.getHttpStatus());
 	}
 
 	/**
 	 * {@inheritDoc}
-	 * Calls {@link WebhookHandler#onMessage(Notification)}
+	 * Calls {@link WebhookHandler#onMessage(Notification.Delta)}
 	 */
 	@Override
 	public void onMessage(String message) {
@@ -78,7 +77,9 @@ public class Tunnel extends WebSocketClient {
 
 		// Parse notification from JSON body and call onMessage callback
 		Notification notification = Notification.parseNotification(jsonBody);
-		webhookHandler.onMessage(notification);
+		for(Notification.Delta delta : notification.getDeltas()) {
+			webhookHandler.onMessage(delta);
+		}
 	}
 
 	/**
@@ -87,7 +88,6 @@ public class Tunnel extends WebSocketClient {
 	 */
 	@Override
 	public void onClose(int code, String reason, boolean remote) {
-		log.trace("Closing websocket connection");
 		webhookHandler.onClose(code, reason, remote);
 	}
 
@@ -97,7 +97,6 @@ public class Tunnel extends WebSocketClient {
 	 */
 	@Override
 	public void onError(Exception ex) {
-		log.trace("Error encountered during websocket connection");
 		webhookHandler.onError(ex);
 	}
 
@@ -184,9 +183,18 @@ public class Tunnel extends WebSocketClient {
 	 * An interface for implementing classes to handle events from the {@link Tunnel}
 	 */
 	public interface WebhookHandler {
-		void onOpen(short httpStatusCode);
-		void onClose(int code, String reason, boolean remote);
-		void onMessage(Notification notification);
-		void onError(Exception ex);
+		void onMessage(Notification.Delta delta);
+
+		default void onOpen(short httpStatusCode) {
+			log.trace("Opening websocket connection. Code: {}", httpStatusCode);
+		}
+
+		default void onClose(int code, String reason, boolean remote) {
+			log.trace("Closing websocket connection. Code: {}, Reason: {}, Remote: {}", code, reason, remote);
+		}
+
+		default void onError(Exception ex) {
+			log.error("Error encountered during websocket connection. Exception: {}", ex.getMessage());
+		}
 	}
 }