+++ CHANGELOG.md
@@ -8,9 +8,14 @@ This section contains changes that have been committed but not yet released.
 
 ### Changed
 
+* Provide default implementations for `WebhookHandler` methods `onOpen`, `onClose`, and `onError`
+
 ### Deprecated
 
 ### Fixed
+
+* Fix `Tunnel.onMessage` not emitting individual deltas
+* Change `java-websocket` dependency to `api` configuration
 
 ### Removed